pyproject.toml
[project]
name = "loyalty-cards"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "python-barcode>=0.15.1",
    "qrcode>=8.2",
]
